airflow/requirements.txt
joblib==1.2.0
lightgbm==3.3.5
numpy==1.22.4
pandas==1.5.3
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pyaml==21.10.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
scikit-learn==1.2.1
scikit-optimize==0.9.0
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
xgboost==1.7.4
tensorflow==2.11.0
keras-tuner==1.3.2
python-dotenv==1.0.0